Adjust test code so we no longer rely on dtype for grid.

+++ src/funmaze/graph/grid.py
@@ -9,8 +9,8 @@ from funmaze.graph import IGraph
 GridNode = TypeVar("GridNode", bound=np.generic)
 
 
-def grid_sequential(shape: tuple[int, ...]) -> npt.NDArray[np.uint]:
-    seq = np.arange(0, np.product(shape), dtype=np.uint)
+def grid_sequential(shape: tuple[int, ...]) -> npt.NDArray[np.int_]:
+    seq = np.arange(0, np.product(shape), dtype=np.int_)
     return seq.reshape(shape)
